pd2DTransform: the latest addition to the pd2D family
The class formerly known as pdGraphicsMatrix has been cleaned up and
greatly expanded, and it now behaves like the rest of the pd2D library.
Paths in particular use a lot of transform operations, so this was a
necessary step on the way toward completing the (incredibly tedious)
pd2DPath class.