Large patch to improve dark theme behavior
There are still a number of dialogs that need touch-up (like the Levels
dialog, which uses some custom UI elements), but I'm now running the
dark theme full-time locally, and gradually fixing problem spots as they
arrive.

The bulk of the work here is fixing dialog startup behavior to apply all
themes before the dialog first hits the screen, which solves flickering
issues when a dark theme is in use.