Continued dark theme improvements
More of the same: making sure initial dialog rendering happens *before*
the form is visible, so we don't get nasty flicker when a dark theme is
active (and the default window background of blinding white is swapped
with a darker color).

Some dialogs may require special treatment due to the way VB arbitrary
decides to show windows (without any obvious way of trapping the show
event, short of subclassing).  I'm looking into this.