New File > Batch > Repair option
After a long holiday, it's good to be back to work on PD!

This new feature is kind of random, but hopefully useful to others.
While on holiday, a family member used a thumb drive full of photos in a
sketchy Internet cafe, and the drive became a corrupted mess.  He ran
good ol' scandisk on the drive, which generated a list of generic
"FILE0001.chk", "FILE0002.chk" type files inside a "FOUND.000" folder.
Some of these files are valid images (mostly JPEGs).  Many are not.
Manually testing and renaming each file would be a nightmare, and he
asked if I could help.

So PD now has a batch recovery tool.  File > Batch > Repair is a
cleaned-up version of a quick script I wrote to solve his problem.  It
allows the user to select a source folder (with or without subfolder
recursion), then have PD test each file in the folder.  PD runs some
basic heuristics to try and identify valid images, and if one is found,
the file is copied to a "repaired" folder and renamed according to the
file's actual type.

I realize this is a pretty basic "repair" option, but in the future, it
can always be expanded.