Start work on theme resource editor
This is one of the last big theming items still on the to-do list.
Separating theme resources (like UI images) into their own custom
resource file, which can be updated independently of PD's exe, will be a
huge improvement over the current system.  It will also make it much
easier for the program to do things like auto-select the best image
resource for a given screen DPI.