Start overhauling the Batch Process wizard
As part of migrating PD to its own custom resource file format, I'm
gonna be bulk-optimizing a ton of UI images (mostly PNGs).  The current
batch processor is a mess, so it's time to get it fixed up, and
integrate support for all of PD's new image export features.

Note that this is a WIP, so you *still* need to avoid the batch
processor for now.  I'll remove the warning window when the dialog has
been thoroughly overhauled and tested.