Batch processor: continue overhaul
The "add files to batch list" panel is being completely reworked.  Given
the huge feature set of a standard common dialog window (the full power
of Windows Explorer, basically), it doesn't make sense for us to try and
manually mimic that level of functionality.

Instead, the new dialog is built around adding images and/or entire
folders from standard Windows dialogs.  The ability to add whole folders
is much more elegant than the old solution, anyway, and this approach
significantly streamlines the dialog, and greatly improves performance
as the process list grows.

(Note that not all options on the new panel are implemented yet.)