Batch processor: overhaul export setting UI
One of the major goals for 7.0 is to use the same export settings UI
across both Save/SaveAs in the main program, and also the batch
processor.  This allows the user to fine-tune their batch processor
export settings with the same detail as a regular save operation
(including metadata handling), which is a great improvement.  It also
spares me a ton of redundant UI and translation work.

I have yet to actually test the new implementation, but export setting
dialogs are now correctly raised, and all dialogs support "no preview"
mode - a necessity during batch processing, as there is no base image to
preview!