Workaround for FreeImage handling of some 32-bpp PNGs
FreeImage will sometimes report 32-bpp PNGs with opaque alpha channels
as *not* transparent.  This is confusing as colloquially, that may be
true, but the images are still in RGBA format, so you need to use 32-bpp
per-pixel offsets.  PD wasn't doing this, and it was leading to
problematic handling of 32-bpp images with embedded ICC profiles (as the
ICC engine was processing the images as RGB, not RGBA, on the assumption
that there was no transparency data in the image; this led to mismatched
byte ordering and all kinds of weird display issues).

Anyway, it's simple enough to catch this rare case and work around it,
but I'm concerned by the way FreeImage is reporting this.  I'll do some
extra homework to make sure nothing else relies on their IsTransparent()
function for correct behavior.