Fix the way metadata export preferences work
Metadata export preferences now work the same way as all other export
settings.  They are stored on a per-format basis, with last-used
settings automatically saved/loaded on each invocation.

This means there is now just one metadata export preference in the Tools
> Options dialog: marking PD as the last-used software.  (Note that even
if this setting is enabled, it will be ignored if the "anonymize
metadata" setting is clicked at export time.)