Start total overhaul of primary layer list UI
This old hunk of junk has not yet been integrated into the new theming
engine.  It is also not really high-DPI aware, it's drawn using a bunch
of unsafe subclassing, and it uses things like hover elements which are
not really ideal for our modern touch-centric world.

So it's time for an overhaul.  Among other things, the goal is to make
the layer list box:

1) Themable
2) High-DPI aware
3) Touch-friendly
4) Make better use of space
5) Less crash-prone (there's a lot of unsafe subclassing going on now)
6) Less resource-intensive (it uses a ton of custom graphics in dumb
ways)
7) More responsive (*very* important for paint tools)

A number of these items have already been accomplished in this initial
commit.  The bulk of the control has been migrated to a new user
control, rendering has been rewritten against PD's master user control
support class, and theming has been enabled.

Next up is restoring scroll behavior (there is no embedded scroll bar at
present), and redesigning the individual layer blocks to not rely on
hover behavior.