Layer UI: overhaul completely
Building off the last commit, there are now no hover-only elements in
the layer list box.  This makes that area much more touch-friendly,
while also reducing visual clutter.  (Note that the plan is to move the
removed UI items to a right-click menu; that is still TODO.)

The new layer box layout is much closer to the competition, with the
visibility toggle moved next to the edit area, thumbnails slightly
shrunk, and much more room for the layer name.  Clickable regions have
been enlarged, and the visibility toggle in particular now received a
hover border, to let users know that it's clickable (in keeping with all
other clickable UI elements in the project).

The last thing left to solve here is reinstating the scroll bar for
images with many layers.