pdTooltip: implement rendering
Tooltip positioning still needs to be optimized (for example, no checks
are currently performed to prevent displaying off-screen), but tooltip
window management and rendering are finally working well.  (Translations
are not yet applied, but that's a quick fix.)

Also note that so far I've only tested Win 10.  Things may be weird on
XP and/or Vista; testing them is TODO.

This was the single largest remaining item on the "reduce user and GDI
object" checklist.  Standard Windows tooltips leak like a sieve when
intermixed with self-subclassing VB methods (because tooltips also
subclass their target objects, and their common-control-based
subclassing does not play nicely with Paul Caton's approach).  Just the
*assignment* of a tooltip resulted in the creation of a user object and
two GDI objects, and those resources could potentially leak any time a
tooltip was displayed or changed.  (As noted in previous commits,
changes to the Title property of a tooltip would cause an additional
*4-6* GDI objects to leak!)

Thus, switching to a custom tooltip implementation is a big deal.  By my
accounting, this cuts cold-start user objects by ~150 (662 to 514), and
GDI objects by nearly ~200 (1278 to 1081).  This is a lower user object
count than PD 6.6 (626) and a GDI count that is nearly *half* of what it
was in 6.6 (1081 vs 1981).  And all this despite moving to a fully
owner-drawn UI, with double-buffering for all display elements.