Master UI class: free DCs after controls raise a redraw request
There's no penalty to doing this, as DCs are autogenerated as needed.
There is a nice benefit, however, if a control requests a redraw, but
the system doesn't generate an immediate paint event (because the
control is hidden, for example).  This is particularly useful at startup
time, as many controls are initialized, but not ultimately displayed
on-screen.

On a cold start, this reduces GDI object count by 245.  The resulting
count is less than not just PD 6.6, but PD 6.4 as well.