Minor bugfixes
This commit fixes minor rendering issues of button tooltips (the font
object needs to be manually reset to left-alignment), and the main
canvas at program shut down (it was incorrectly suspending rendering at
the wrong shutdown phase).

Also, I've confirmed that the source of user object leaking at shutdown
is the self-subclassing class.  I've been meaning to switch to the
standard common-control-based subclasser, and this provides a good
reason to do it sooner rather than later.

Such a switch first requires a rewrite of the color selection dialog
(which currently uses a bunch of outdated control approaches,
incompatible with the new subclasser), but that's not a bad thing as I
needed to redesign the color selector anyway to make it theme- and
high-DPI-compatible.

In the meantime, this commit makes some additional minor subclassing
changes to prepare for the rewrite.