Color selector: switch to pdColorWheel
Eventually, the goal is to make both sides of this dialog user-swappable
(with support for swatches and other color selection methods), but for
now, I just want to migrate everything to a theme- and
high-DPI-compatible implementation.

Converting the left-hand side to an existing control is an easy way to
jumpstart this.  Among other benefits, redraw speed is vastly improved
over the old system.