pdSlider: new visual style property
Sometimes, pdSlider is used to select a particular color from a given
range of colors.  (Often, the range of colors is a smooth gradient, and
the slider selects a point along that gradient.)  An automatic
"gradient" visual property lets the caller render a background gradient,
to make the control's purpose very clear.

Unfortunately, the current visual design uses the slider knob to overlay
the currently selected position, which obscures the color underneath.
In most cases this isn't a big deal (as color specificity isn't really
the point), but in the actual color selection dialog, it's a problem.
Color specificity is the whole point!

To that end, this commit adds a new visual style option for the slider,
where instead of rendering the knob as a filled circle, the knob becomes
a hollow rectangle.  This allows the currently selected position to be
seen clearly.

I want to do a bit more tweaking to the slider's background for this
style, but this commit at least works (and the slider is still fully
functional).