pdDIB: heavily optimize alpha channel transfers
For some UI elements (like pdSliders with color gradients in their
tracks), we steal alpha values from a reference DIB and apply them
on-the-fly to some other DIB.  This greatly simplifies rendering, as we
only have to create the alpha mask once.

The routine that handles the transfer has been heavily optimized, and if
the reference alpha mask only deals in binary alpha channels (meaning
fully opaque or fully transparent pixels only), it's *extra* fast.