Color selection dialog: overhaul HSV sliders
The new HSV sliders are standard pdSlider controls, and the new
owner-draw mode is used to render custom HSV gradients on-the-fly.  This
means the color selection dialog is *finally* usable on a dark theme.
(Yay!)

Also, performance of the dialog is now much better, thanks to all the
recent pdSlider improvements.

Next up for this dialog?  A new custom-built control for the "recent
colors" area.  Besides obeying the same UI behavior as everything else
in the program, I want a control like this for use in other places, to
make it easy to display an owner-drawn "history" of anything in the
program.