pdNewOld: a new control for side-by-side comparisons...
...between any arbitrary "new" and "old" element.  (The elements
themselves are owner-drawn.)

At present, this control is used by the color selection dialog to
present the original and newly selected colors side-by-side.  The color
dialog does not yet color-manage these results.  (That's still "to-do".)

Also included in this commit is an update for pdColorWheel, which
improves responsiveness during user interaction or color change
requests.