Color selector: overhaul "color history" area with new pdHistory control
First, an apology for a lack of recent PD updates.  I'm in the midst of
a large project that has a mid-November deadline, and it's imperative
that I finish on time.  This is consuming all my free time, so PD
updates may continue to be sparse over the next month.  Sorry!

That said, I've finally wrapped work on a great new pdHistory control,
which manages a "recently used" list for any arbitrary object type.  At
present, it's used by the color selection dialog to track the user's
recently used colors.

pdHistory automatically handles the history stack, and it simply raises
rendering events so that the owner can paint the history however they
like (with pdHistory managing things like mouse behavior and layout).
This allows for much more complicated history presentations; on the
color selection dialog, for example, we can now display multiple rows
and columns of previously used colors while using the same visual
language as other PD controls.

As part of this overhaul, I've also reworked the default color history
palette.  The new default history is much prettier.

This is an important commit because the color selector was the last
place in PhotoDemon where problematic old mouse subclassing techniques
were used.  With that dialog successfully overhauled, I can now rework
the program's core mouse handler to finalize input preparations for
paint tools.  (In particular, I can finally make it so that the main
canvas supports high-resolution mouse input, while other UI elements do
not.  At present, this is a program-wide behavior - not good!)