Central mouse handler: total overhaul
As discussed in 8fbd17f320dcaaafd34248ab65aa3331a2f301ec, recent changes
to the color selection dialog have finally allowed me to do a
large-scale overhaul of pdInputMouse, the primary mouse handler for the
program.

As part of this overhaul, the class is now much leaner, lighter, and
more reliable.  It is also primed for moving to a new subclassing
solution in the near future.