Fix a number of Undo/Redo issues with non-destructive properties
Non-destructive properties are a real PITA when it comes to Undo/Redo.
When toggling something like layer visibility, we don't want to create
Undo entries for every change - for example, if the user toggles a
layer's visibility "on" then "off", we don't want to create an Undo for
that, as the image is back to its original state.

This commit should fix bothersome issues where Undo/Redo data would
seemingly be created out of nowhere, because the system was overly
aggressive about trying to track changes to non-destructive properties.