New merge pipeline for paint operations
Building off the previous commit, the final "merge" of a paint brush
onto a target layer is now handled by a specialized merge function
(instead of the regular MergeLayers() function, which uses different
order rules).

This means that previewed brush strokes and final brush strokes are now
identical - yay!

With this, I believe the basic paintbrush is pretty much free of major
bugs.  Performance is still not ideal, but I've got a gameplan for
fixing it.  Now if only I can find the time...