pdSlider: new exponential, logarithmic modes
Some sliders cover an enormous range of values.  If users are likely to
spend most of their time in a certain subset of values (e.g. a slider
that goes from 1 to 10,000, but values 1-100 are most likely), this
makes the slider much more usable.

There are many sliders in the project that could make use of this
option, but for now, I've only updated the Paintbrush Size slider.  I'll
get to others in due time.

(Also, note that this new setting *only works for sliders with minimum
values > 0*.  This is by design.)