Use ExifTool to extract ICC profiles from formats FreeImage doesn't support
Also, a generic function now exists to use ExifTool to extract *any*
embedded ICC profile out to a standalone file.

Because metadata is extracted asynchronously, the base image may appear
on-screen before ExifTool pulls the ICC data for us.  This may cause the
image to "flicker" as its correct profile is applied.  I don't consider
this a problem, as the cost of fixing it is waiting for ExifTool
synchronously, which is simply not feasible from a performance
standpoint.