pdLayer: lay the framework for migrating ICC profile management here
With this, it's now time to do something scary: start work on an upgrade
to the PDI file format.  I'm going to do what I should have done in the
first place (sigh) and use the .zip file approach of sticking the
directory at the *end* of the file.  This lets us write compressed data
out to file immediately, without the need for an in-memory cache, and
when we're done we just stick our finished directory after all the
compressed data.

Also, I'm looking at moving from zLib to Facebook's zstd library
(https://github.com/facebook/zstd) which provides compression levels
nearly identical to zLib, but with significantly faster compression and
decompression time.  This would be a large win for us, particularly when
writing Undo/Redo data.

Anyway, this all relates to ICC profiles because a new chunk is needed
inside PDI files for storing all ICC data for all layers (and
potentially the image's working space, too).  But I don't want to
implement this until the new PDI implementation is finished.