PDI file format: roll over various engines to v2
v2 of the PDI format is going to be built on v2 of the pdPackager class.
This format will break backwards compatibility with v1 in a number of
significant ways.  It will also all be handled transparently to the end
user, so there's nothing here for users to be concerned with.

On my end, v2 is going to be a significant improvement for both memory
usage and performance.  pdPackager2 is going to completely do away with
the need to maintain a running copy of the archive in RAM.  Instead, it
will write data immediately out to file as it receives it, allowing us
to work with much larger files and with improved asynchronous behavior
(thanks to everything being built around memory-mapped files from the
start).

Also new to v2 is support for the new zstd compression library, which
will give us much better compression performance at identical or better
compression ratios than v1.  Integrating zstd is still "to-do", however.