New unified compression interface
The new Compression module makes it trivial to perform compression and
decompression tasks against any backend.  Simply call the standard
compression/decompression functions, specifying the compression engine
you want, and the module silently takes care of the rest.

This greatly simplifies the design of pdPackager2, and should allow me
to remove a lot more redundant compression code from throughout the
project.  It also makes it much easier to add new compression libraries
to the project, as it's just the addition of a new compression engine
enum to the central Compression module; everywhere else in the project
inherits that enum "automagically".

As part of this change, the default compression engine for internal PD
bits is now zstd.  This accelerates a number of previously slow
processes, but don't get too comfortable, because they're soon going to
get even faster...