Compression engine: add support for lz4_hc
As noted in the code, lz4_hc supports comes "free" with the lz4 library.
lz4_hc is a high compression variant of lz4, using an identical block
and frame format (so there's no unique decompression function), but with
a much more aggressive search strategy during compression.  This results
in better compression results but much slower compression performance.
Note that decompression speed is identical to regular lz4, so lz4_hc is
a good candidate when there is plenty of time available for compression,
but decompression performance is at a premium.  I'm considering using it
for PD's new resource file format, for example.

Anyway, lz4_hc compression is now supported both throughout PD and
within pdPackages specifically.  Note that the pdPackages spec isn't
affected by this - the file flags still just use LZ4 as a descriptor,
since only the LZ4 decompressor is required.