pdPackager: implement file-backed mode when reading packages from disk
This (significant) change means that a pdPackage file (like PDI images)
no longer needs to be copied into memory when you are reading it.
Instead, the short file header is verified, then the file *directory* is
copied into memory.  The file data itself lives on disk, and is only
accessed as individual nodes are retrieved.

This greatly cuts the amount of memory required to read PDI files, while
performance remains roughly the same.

Note that most of the actual work of this occurred inside the pdStream
class, which is responsible for invisibly managing pointers in either
memory-backed or file-backed mode.

Next up is providing the same option when writing PDI files.