pdPackager: implement "write as you go" file support
This greatly cuts down on memory usage, but there's a hefty performance
hit to go with it (presumably because the file size isn't known in
advance, so new allocations may be happening on every write).

Because of this performance hit, I'm only enabling the file-backed mode
for saving primary PDI files.  Undo/Redo still uses a memory cache
because it's a good 2x faster.

I'll keep investigating ways to improve performance, but for now, this
feels like a reasonable solution.