Compression engine: add support for Windows Compression API
The usefulness of this API is questionable (especially because it's
limited to Win 8+), but it's only a few lines of code, and it gives me
some options for testing.  On my development PCs, none of the MS engines
outperform our existing 3rd-party solutions, which isn't exactly
surprising, but I'll run additional tests on other PCs to confirm.

So as-is, this is just "FYI", because it's still much faster for us to
use zstd and lz4, but perhaps others will find it useful.

Also included in this commit are new functions for retrieving each
library's max/min/default compression levels.  This reduces guesswork
when validating input.