Paintbrush: improve cursor accuracy and behavior
The central cross is now removed while the mouse is down.  (Its main
benefit is seeing which pixel you'll target before clicking, so there's
no reason to keep it around once the mouse is down.)  I may also look at
a property for hiding the outline entirely while the mouse is down; for
basic brushes, this may be the preferred behavior.