Brush engine: implement line rasterizer
Actually, there are three line rasterizers in this placeholder commit.
After a bunch of testing, I've decided to go with a voxel traversal
technique based on this lovely paper:
http://www.cse.yorku.ca/~amana/research/grid.pdf.  I prefer this
technique over something like Bresenham because it calculates every
pixel intersected by the current line, rather than just the "pixel with
the largest intersection at each point".  This produces better coverage,
particularly when variable brush spacing is taken into account (as it
must be for increasingly large brushes).

There's some fun debug code in this for the curious who want to test the
brush engine against any arbitrary source image file, but if you'll wait
just a bit, I'll shortly be dropping in a true "brush hardness"
renderer, which will allow anyone to play with the new brush engine...