Brush engine: improve adaptive framerate approach
This commit is designed to improve UI responsiveness at both the
high-end and low-end.  (Middling performance was already handled well by
the old algorithm.)

The brush engine now keeps a running tally of frame rendering times.
Unlike before, if background brush renders occur fast enough, the
viewport synchronizer is now allowed to run at an uncapped framerate,
basically dumping updates to the screen as fast as the brush engine can
supply them.

If the "uncapped" approach starts to dip into problematic territory (<
15 fps, at present), we start dropping viewport synchronizations to free
up more time for the background brush engine to keep pace with incoming
mouse events.  Frame dropping increases in severity until a minimum
threshold of 2 fps or less is hit.  At that point, we enforce viewport
synchronizations after every update, which is deeply unpleasant but the
best we can do.  (Note that a scenario like this should never be
obtainable on a modern PC, unless you do something crazy like make an
enormous brush then manually set brush spacing to its minimum level.)

Note that regardless of viewport synchronization rate, the paint cursor
is always updated in sync with received mouse events, so the only
penalty from a UI standpoint is a slight delay as the underlying stroke
periodically "catches up" to the cursor position.

Of course, at common brush and image sizes, none of this should matter
because the brush engine is fast enough to cope - but as we add new
brushes with more intense modeling requirements, this adaptive framerate
stuff could prove useful.