Load specific tool panels on-demand...
...instead of pre-caching them at start-up time.

There are pros and cons to both systems.  The biggest "pro" of
pre-loading panels is that they subsequently load very, very quickly
when the user switches tools.  The biggest "con" is that these are
large, dense panels with a lot of UI elements, and loading them takes
time and resources - and those resources stick around for the lifetime
of the program.

If we load tool panels on-demand, PD's cold startup resource count drops
from 559 to 187 user objects, and 909 to 457 GDI objects.  The amount of
memory used is also reduced, while startup time is improved.

These are significant improvements, and if a user isn't going to use a
particular tool in a session, I don't like wasting energy loading the
corresponding panel in advance.

If this proves too aggressive (e.g. users are unhappy/confused by slight
delays in tool switching), I can look at restoring the former behavior,
or against my better judgment, perhaps adding a toggle in the Options
panel.  I also plan on breaking out an old PC for testing in the coming
days (I've got lots of new features that need to be checked on XP) and
I'll make sure the delays aren't too egregious.  (And if they are, maybe
I can find ways to optimize various user control initializations.)