About dialog: total redesign
I don't like spending time on non-functional stuff like this, but the
old About dialog was a mess.  Besides using an animated credit scroller
(ugh), it also used a bunch of custom graphics that consumed an
unnecessary amount of storage.

The new dialog is much more minimalist, and it even contains
(potentially) useful information!  I'm sure it could be worked into
something fancier, but that's not where I want to spend my time, so this
will have to do for now.