Adjustments> Levels tool: modernize UI
The "level node" interface bits are now dynamically rendered at
run-time, using theme settings, instead of pulling static images from
the resource file.  This means the dialog now adheres to the same UI
rules as the rest of the program (finally!).

I've also removed a number of other unused resources, and as a result,
.exe size continues to drop.