Effects > Stylize > Outline
Since I have this code available, I may as well expose it to the user.

At present, the function only renders the *first* discovered outline on
a given image.  It works best for contiguous images on a transparent
background, and there is also an option for solid-colored background
detection.  I could rework it to detect *all* contiguous outlines in an
image, but this makes it quite a bit slower... so it'll have to wait.