Update generator+patcher: add support for new theming data
Eventually I'll build the theme files directly into PD's executable
file, so it can extract them as a failsafe if they go missing or become
corrupted.  For now, however, the files are still in flux so I don't
want to commit them just yet.  (In other words, don't delete or screw
with the /App/PhotoDemon/Themes folder - it's essential to proper
program operation.)

That said, PD's update builder and patcher have been updated to create
and patch files in the /Themes folder.  This should fix problems with
broken auto-upgrades from past versions.

(Thank you to ChenLin for mentioning this problem to me.)