Big drop: lots of new 256-color / palette code
PhotoDemon is (and always will be) focused on editing photos.  But that
doesn't mean we can't do some cool things with palettes and indexed
color modes.  (For example, there are a number of interesting features,
adjustments, effects, and paint ops that lean on things like custom
palettes.)

To that end, I've dropped a bunch of palette-related code into PD.  At
present, this code is not exposed to the user in meaningful ways.
Instead, it is a large collection of algorithms, modules, and classes
built around palette generation and application.  Effects and
adjustments that use these tools are coming soon, alongside other
feature-centric 7.0 development.

Included in this commit are:

1) A new "ResizeDIBByPixelCount" function.  This function resizes a
target image to a specific number of pixels (e.g. "resize to 50,000
pixels"), while preserving aspect ratio.  This is extremely helpful for
things like image heuristics, because heuristics don't need a full copy
of the image.  A small copy still provides all the information we need
(when building things like histograms or color maps), and this function
lets those heuristics run at a constant time, independent of image size.

2) A new "GetOptimizedPalette()" function.  Given an arbitrary source
image, this function returns an optimal palette of that image on the
range [2, 256] colors (as specified by the caller).  This function is a
custom-built median cut-like implementation, using pure VB6, and it's as
fast (or faster) then the existing color quantizers we've been using
from the FreeImage library.

The main reason I wanted to custom-build my own quantizer is because
FreeImage's quantizers require a lot of memory, they're not especially
fast, they can't just return a palette (e.g. they *always* apply the
palette to the image) and the library as a whole is prone to problematic
edge-cases that aren't well-tested.

Also, like most other features in the program, we can generally produce
ones than free 3rd-party libraries.  (For example, the current quantizer
runs in RGB space, but it would be trivial to add a Lab version for
improved results.)

Anyway, the quantizer is well-documented and straightforward in
operation.

3) Four (!!!) different "ApplyPaletteToImage" functions.  Not all of
these functions may remain in the project, but I wanted to commit them
all in case a revert is needed in the future.

As it turns out, applying a palette to an image is actually much more
time-consuming than generating the damn palette in the first place.  As
such, I've tested a number of different prototypes over the past year,
with a few interesting discoveries along the way.

In order of performance (slowest to fastest), here are the ApplyPalette
algorithms now available to PD:

1) A naive, perfect matching function that searches the entire palette
for a "best" match against each pixel.  As you'd expect, this function
is slow, but it's useful because we can modify the color-matching
algorithm to something much better than plain old Euclidean distance.
(For example, it currently uses a weighted comparison where each color
component is weighted by its sensitivity to the human eye.)

This function will always yield the best results, but it'll take awhile
to get there.

2) A WAPI-optimized version of (1) above, using standard Euclidean
distance matching for colors.  For whatever reason, GDI's
GetNearestPaletteIndex() function is much faster than an identical
function in VB, so this cuts the time of our naive function by about
75%.  (It's also equally fast in the IDE or compiled, which is nice
during testing phases.)

3) A lossy hash-table lookup.  Colors are sorted by their distance from
black, than organized into [n] hash buckets (where the caller controls
[n]).  More buckets = faster performance but lower quality.  At its
default setting, this is roughly 2x faster than an exhaustive match
using GDI, while the resulting image is very close quality-wise.

Hash tables are particularly useful if applying a palette with
dithering, because dithering automatically compensates for the small
errors due to hash table misses.  It is also unique in giving the caller
some control over the performance vs accuracy trade-off.

4) A self-pruning octree color lookup.  Octrees can also be used for
palette generation (and I may actually implement this in the future,
although our median-cut system will likely be better quality-wise), but
here, I'm purely interested in using them to accelerate the
color-matching process.  A self-pruning system means that colors are
always placed in the highest-possible tree depth, making color lookups
about as fast as you can physically get in VB (outside of VB-specific
performance tricks).  This system is roughly 4x faster than the
"perfect" GDI solution, but note that it is *not* guaranteed to return
ideal matches.  (Colors are placed in the tree by bit-order, so colors
that lie very near bit boundaries don't match well - e.g. RGB(127, 127,
127) will map to RGB(0, 0, 0) instead of RGB(128,128,128) because it
shares a top-most bit with the former, despite the latter being closer).

I have some ideas for mitigating the imperfect color-matching nature of
octrees, but as with hash-tables, the lossy nature is not a very big
deal when dithering, as dithering minimizes the visual impact.  (In
fact, the end result during dithering can sometimes be *better* if
colors are matched imperfectly, as you escape some of the patterned-ness
inherent to dithering.)

So that's an overview what this commit contains.  Part of the reason I
wanted to commit it now is that with paint tool development in full
swing, the color selector would greatly benefit from a "generate a
palette with colors from my current image" feature.  We can now do that
with just a few lines of (very fast) code.

Note that this commit is also a prerequisite to proper swatch support,
which has been in the pipeline for a very long time.