Effects > Stylize > Palettize: many new features
(Note: this effect was formerly the Image > Indexed color... menu, but
as explained in a previous commit, it no longer makes sense there.)

- Dithering is now supported across all adaptive quantization methods.
11 different options, including both ordered (fixed) and error-diffusion
methods are available.
- An option to precisely preserve white and black in the final palette
is now available.  Note that white and/or black are only preserved if
they exist in the original image.  (Otherwise, we'd be wasting palette
entries on them.)
- Reduced color bleed is now an available option.
- Minor bugfixes to the pdMedianCut class.