PNG export: improved color/transparency format settings
This presentation makes much more sense.  It will also provide a
much-better template for other dialogs, while greatly reducing the
amount of bug-testing I have to do (as export dialogs can now lean on a
single user-control to manage color depth settings).

Next up, rewriting the function that takes the color depth settings and
actually produces a matching image from them.  At present, this function
leans on a horrifying hybrid of custom PD code and the FreeImage library
to perform the transition.  We'll be reworking this to lean on PD code
only, with FreeImage only being used for the final export.  (Again, this
should *greatly* reduce bugginess of export dialogs.)