Many small performance improvements
No "huge" performance improvements here, but this commit is an effort to
make many small improvements across various parts of the program (with a
special emphasis on functions hit during program startup).

The biggest changes relate to UI theming, particularly the way object
color tables are constructed and image resources are retrieved.   I've
also fixed some intermittent bugs with the master language translation
class (which caused its caching behavior to sometimes fail), improved
theming speed on the scroll bar class (which is important as scroll bars
are embedded in tons of other controls), and fixed a number of other
obnoxious intermittent performance issues.